Renamed the ModificationExtension folder to ModificationExtensions (added s).
